requirements.txt
alembic==1.8.1
bcrypt==4.0.1
black==22.10.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
dnspython==1.16.0
email-validator==1.3.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-CKEditor==0.4.6
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathlib==1.0.1
pathspec==0.10.2
platformdirs==2.6.0
psycopg2-binary==2.9.5
pycparser==2.21
python-decouple==3.6
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
Werkzeug==2.2.2
wincertstore==0.2
WTForms==3.0.1
